Move current version file and fix relative import
