+++ emmet-api/setup.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from setuptools import find_namespace_packages, setup
-from _version import __version__ as fallback_version
+from emmet.api._version import __version__ as fallback_version
 
 if "+" in fallback_version:
     fallback_version = fallback_version.split("+")[0]
