+++ emmet-builders/setup.py
@@ -1,5 +1,5 @@
 from setuptools import find_namespace_packages, setup
-from _version import __version__ as fallback_version
+from emmet.builders._version import __version__ as fallback_version
 
 if "+" in fallback_version:
     fallback_version = fallback_version.split("+")[0]
