--- emmet-core/_version.py
@@ -1,1 +1,0 @@
-__version__ = "0.38.1"
