+++ emmet-core/emmet/core/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.21.13"
+__version__ = "0.38.1"
